requirements.txt
# MCP and HTTP
mcp>=1.0.0
httpx>=0.27.0

# AWS Bedrock
boto3>=1.34.0
botocore>=1.34.0

# Data validation and parsing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0

# YAML parsing for input
PyYAML>=6.0.0

# Utilities
rich>=13.0.0

